refactor(test): split marker-schema-migration tests out of test_review_marker_companion_verdicts.py
The file crossed 300 lines after adding schema-4/GLM migration coverage
(anti-ratchet bloat gate flagged the new crossing). Extracted the three
historical-roster/unrecognized-roster tests into a sibling module rather
than raising a baseline cap, per this repo's bloat-vs-long-code convention.
