+++ shared/tests/test_review_marker_companion_verdicts.py
@@ -136,37 +136,6 @@ def test_companion_writes_and_repairs_from_the_recorded_verdicts(tmp_path):
     assert json.loads(Path(repaired[0]).read_text(encoding="utf-8"))["verdicts"] == current
 
 
-def test_companion_writes_and_repairs_a_historical_deepseek_envelope_as_schema_3(tmp_path):
-    # Regression: write_markers used to fall back to MARKER_SCHEMA (4) for any
-    # non-gemini roster, so repairing a pre-swap deepseek/openai record wrote
-    # a schema-4 marker carrying a deepseek roster - which evaluate_review_state
-    # then blocks, since schema 4 requires glm/openai. A historical record must
-    # stay both readable AND writable/repairable as schema 3.
-    historical = {"deepseek": "approve", "openai": "revise"}
-    paths = write_markers(
-        tmp_path, "run-1", "plan", marker_status="completed",
-        record_status="completed", findings_count=0, verdicts=historical,
-    )
-    marker = json.loads(Path(paths[0]).read_text(encoding="utf-8"))
-    assert marker["marker_schema"] == 3
-    assert marker["verdicts"] == historical
-    state, _ = evaluate_review_state(marker)
-    assert state != STATE_BLOCK
-
-    record = upsert_review(
-        new_record("run-1"),
-        make_entry("plan", STATUS_COMPLETED, verdicts=historical),
-    )
-    write_record(tmp_path, "run-1", record)
-    Path(paths[0]).unlink()
-    repaired = repair_markers(
-        tmp_path, "run-1", "plan", marker_status="completed"
-    )
-    repaired_marker = json.loads(Path(repaired[0]).read_text(encoding="utf-8"))
-    assert repaired_marker["marker_schema"] == 3
-    assert repaired_marker["verdicts"] == historical
-
-
 def test_companion_repairs_the_recorded_operator_resolution(tmp_path):
     verdicts = {"glm": "approve", "openai": "reject"}
     resolution = "Operator accepted OpenAI rejection and corrected the implementation."
@@ -184,32 +153,6 @@ def test_companion_repairs_the_recorded_operator_resolution(tmp_path):
     repaired = repair_markers(tmp_path, "run-1", "plan", marker_status="completed")
     marker = json.loads(Path(repaired[0]).read_text(encoding="utf-8"))
     assert marker["contradiction_resolution"] == resolution
-
-
-def test_companion_writes_a_historical_gemini_envelope_as_schema_2(tmp_path):
-    historical = {"gemini": "approve", "openai": "revise"}
-    paths = write_markers(
-        tmp_path, "run-1", "plan", marker_status="completed",
-        record_status="completed", findings_count=0, verdicts=historical,
-    )
-    marker = json.loads(Path(paths[0]).read_text(encoding="utf-8"))
-    assert marker["marker_schema"] == 2
-    assert marker["verdicts"] == historical
-    state, _ = evaluate_review_state(marker)
-    assert state != STATE_BLOCK
-
-
-def test_write_markers_rejects_an_unrecognized_verdict_roster(tmp_path):
-    # A gateway-route pair (GATEWAY_REVIEWERS = "model-1"/"model-2") or any
-    # other unrecognized reviewer set must fail loudly here rather than fall
-    # through to MARKER_SCHEMA silently, mirroring the fix that now applies
-    # to the historical deepseek/openai roster.
-    with pytest.raises(ReviewRecordError, match="unrecognized reviewer set"):
-        write_markers(
-            tmp_path, "run-1", "plan", marker_status="completed",
-            record_status="completed", findings_count=0,
-            verdicts={"model-1": "approve", "model-2": "approve"},
-        )
 
 
 def test_completed_companion_without_verdicts_is_refused(tmp_path):
